Move prettyCode function into the utility module.
